requirements.txt
pyautogui
selenium
webdriver-manager
Pillow
python-dotenv